+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1
+
+* Improved documentation.
+* Provided `example/list_view.dart`.
+* Provided `example/page_view.dart`.
+
 ## 0.1.0
 
 * Initial Release.